@@ -1,11 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Threading.Tasks;
 using SFA.DAS.Tools.AnalyseErrorQueues.Domain;
 
 namespace SFA.DAS.Tools.AnalyseErrorQueues.Services.DataSinkService
 {
     public interface IDataSink
     {
-        void SinkMessages(string envName, string queueName, IEnumerable<sbMessageModel> messages);
+        Task SinkMessages(string envName, string queueName, IEnumerable<sbMessageModel> messages);
     }
 }
